# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Whatsapp Api",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://official.whatsapp.loc/engine"
        }
    ],
    "paths": {
...
        "/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "fetch all messages",
                "description": "fetch user all messages",
                "operationId": "messages",
                "parameters": [
                    {
                        "name": "messageId",
                        "in": "query",
                        "description": "Message Id to get specific message",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page Number",
                        "required": false
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "Page Size",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns array of messages."
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
...
        "/channels": {
            "get": {
                "tags": [
                    "Channels"
                ],
                "summary": "fetch user channels",
                "description": "fetch user all channels",
                "operationId": "channels",
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            },
            "channel_id": {
                "type": "apiKey",
                "description": "Enter Your Channel ID",
                "name": "CHANNELID",
                "in": "header"
            },
            "channel_token": {
                "type": "apiKey",
                "description": "Enter Your Channel Token",
                "name": "CHANNELTOKEN",
                "in": "header"
            }
        }
    }
}